Extract utility functions

+++ All/install_scripts/nerd-fonts.sh
@@ -1,23 +1,19 @@
 #!/usr/bin/env bash
 
-function setcolor {
-	if [ -x "$(command -v setcolor)"  ]; then
-		$(command -v setcolor) "$@"
-  elif [ -x "${FRESH_LOCAL}/All/bin/setcolor" ]; then
-		"${FRESH_LOCAL}"/All/bin/setcolor "$@"
-  fi
-}
+source "${FRESH_LOCAL:-${HOME}/.dotfiles}/utilities"
 
 ################################################################################
 # Clone/Pull Nerd Fonts Repo
 ################################################################################
+
 NERD_FONT_REPO="${FRESH_PATH}/source/ryanoasis/nerd-fonts"
-mkdir -p "${NERD_FONT_REPO}"
-( \
+if _needs_git_pull "${NERD_FONT_REPO}"; then
 	git -C "${NERD_FONT_REPO}" reset --hard HEAD \
-	&& git -C "${NERD_FONT_REPO}" pull --depth 1 \
-) \
-|| git clone --depth 1 https://github.com/ryanoasis/nerd-fonts.git "${NERD_FONT_REPO}"
+	&& git -C "${NERD_FONT_REPO}" pull --depth 1
+else 
+	mkdir -p "${NERD_FONT_REPO}"
+	git clone --depth 1 https://github.com/ryanoasis/nerd-fonts.git "${NERD_FONT_REPO}"
+fi
 
 ################################################################################
 # Create Custom Monoid Nerd Font
@@ -32,16 +28,20 @@ unzip -u "${BASE_MONOID}.zip"
 popd 1>/dev/null
 
 pushd "${NERD_FONT_REPO}" 1>/dev/null
+local font_options=${BASE_MONOID#Monoid-}
 for font in /tmp/monoid/*.ttf; do
+	local font_weight=${font##*/Monoid-}
+	font_weight=${font_weight%-${font_options}.ttf}
+	
   # Generate Standard Nerd Font
-	echo "$(setcolor -f blue)==>$(setcolor reset) Generating Monoid Nerd Font"
-  ./font-patcher -l -c --careful --outputdir "./patched-fonts/Monoid/complete/" "${font}" &>/dev/null
+	echo "$(setcolor -f blue)==>$(setcolor reset) Generating ${font_weight} Monoid Nerd Font"
+  ./font-patcher -l -c --careful --outputdir "./patched-fonts/Monoid/complete/" "${font}"
   # Generate Mono Nerd Font
-	echo "$(setcolor -f blue)==>$(setcolor reset) Generating Monoid Mono Nerd Font"
-  ./font-patcher -l -c -s --careful --outputdir "./patched-fonts/Monoid/complete/" "${font}" &>/dev/null
+	echo "$(setcolor -f blue)==>$(setcolor reset) Generating ${font_weight} Monoid Mono Nerd Font"
+  ./font-patcher -l -c -s --careful --outputdir "./patched-fonts/Monoid/complete/" "${font}"
   # Generate Windows Nerd Font
-	echo "$(setcolor -f blue)==>$(setcolor reset) Generating Monoid Windows Nerd Font"
-  ./font-patcher -l -c -w --careful --outputdir "./patched-fonts/Monoid/complete/" "${font}" &>/dev/null
+	echo "$(setcolor -f blue)==>$(setcolor reset) Generating ${font_weight} Monoid Windows Nerd Font"
+  ./font-patcher -l -c -w --careful --outputdir "./patched-fonts/Monoid/complete/" "${font}"
 done
 
 ################################################################################
